[refactor] security 디렉토리를 상위 디렉토리로 이동
설명:
- security 디렉토리 내에 JWT와 관련이 없는 클래스들이 포함되었기 때문에 JWT 디렉토리 밖으로 이동했습니다.
