+++ src/main/java/com/coolpeace/global/resolver/AuthJwtPrincipalArgumentResolver.java
@@ -1,7 +1,7 @@
 package com.coolpeace.global.resolver;
 
-import com.coolpeace.global.jwt.security.JwtAuthenticationToken;
-import com.coolpeace.global.jwt.security.MemberPrincipal;
+import com.coolpeace.global.security.JwtAuthenticationToken;
+import com.coolpeace.global.security.MemberPrincipal;
 import org.springframework.core.MethodParameter;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
